Day7 2018/01/17 11:58

+++ DemoAndroid/src/com/zest3k/demoandroid/ChatCardData.java
@@ -7,6 +7,6 @@ public class ChatCardData {
 	public int portraitId;
 	public String name;
 	public String chatContent;
-	public int time;
+	public long time;
 	public int messageCnt;
 }
